Move CoursePath* into Navigation
