+++ ArtemisKit/Sources/Messages/Navigation/PathViewModels.swift
@@ -1,0 +1,50 @@
+//
+//  PathViewModels.swift
+//
+//
+//  Created by Nityananda Zbil on 05.03.24.
+//
+
+import Common
+import Extensions
+import Navigation
+import SharedModels
+import SharedServices
+import SwiftUI
+
+@Observable
+final class ConversationPathViewModel {
+    let path: ConversationPath
+    var conversation: DataState<Conversation>
+
+    private let messagesService: MessagesService
+
+    init(path: ConversationPath, messagesService: MessagesService = MessagesServiceFactory.shared) {
+        self.path = path
+        self.conversation = path.conversation.map(DataState.done) ?? .loading
+
+        self.messagesService = messagesService
+    }
+
+    func loadConversation() async {
+        let result = await messagesService.getConversations(for: path.coursePath.id)
+        self.conversation = result.flatMap { response in
+            if let conversation = response.first(where: { $0.id == path.id }) {
+                return .success(conversation)
+            } else {
+                return .failure(UserFacingError(title: R.string.localizable.conversationNotLoaded()))
+            }
+        }
+    }
+}
+
+@Observable
+final class MessagePathViewModel {
+    let path: MessagePath
+    var message: DataState<BaseMessage>
+
+    init(path: MessagePath) {
+        self.path = path
+        self.message = path.message.wrappedValue
+    }
+}
